Fix 4 failing unit tests: BehaviorTree, Marketplace, TemplateCreator, ScriptBindings
- Fix Parallel node default policy to fail-on-any-failure semantics
- Update TestMarketplace expected catalog count from 15 to 17
- Fix TestTemplateCreator crash (DerivedDirName passed path-unsafe chars as temp dir tag)
- Fix TemplateCreator %zu format strings to %llu for MSVC compatibility
- Fix ScriptBindings InputAction Vec2 -> Vector2 type name

+++ Engine/src/Editor/TemplateCreator.cpp
@@ -53,7 +53,7 @@ bool TemplateCreator::SaveTemplate(const std::string& outputDir, const TemplateM
     }
     static constexpr usize MAX_TEMPLATE_ENTITIES = 10000;
     if (entityCount > MAX_TEMPLATE_ENTITIES) {
-        ENJIN_LOG_ERROR(Editor, "TemplateCreator: entity count %zu exceeds max %zu", entityCount, MAX_TEMPLATE_ENTITIES);
+        ENJIN_LOG_ERROR(Editor, "TemplateCreator: entity count %llu exceeds max %llu", (unsigned long long)entityCount, (unsigned long long)MAX_TEMPLATE_ENTITIES);
         return false;
     }
 
@@ -188,7 +188,7 @@ bool TemplateCreator::LoadTemplate(const std::string& templatePath,
         return false;
     }
 
-    ENJIN_LOG_INFO(Editor, "Template loaded: %s (%zu entities)", outMeta.name.c_str(), result.entities.size());
+    ENJIN_LOG_INFO(Editor, "Template loaded: %s (%llu entities)", outMeta.name.c_str(), (unsigned long long)result.entities.size());
     return true;
 }
